angreal
Module py_logger

A unified logger to bridge python and rust

Functions

register

registers the rust logging interface with the python logging interface.

setup_logging

Modifies the Python logging module to deliver its log messages to the host tracing::Subscriber by default. To achieve this goal, the following changes are made to the module: