angreal::py_logger
Function setup_logging

pub fn setup_logging(py: Python<'_>) -> PyResult<()>

Modifies the Python logging module to deliver its log messages to the host tracing::Subscriber by default. To achieve this goal, the following changes are made to the module:

A new builtin function logging.host_log transcodes logging.LogRecords to tracing::Events. This function is not exported in logging.__all__, as it is not intended to be called directly.

A new class logging.HostHandler provides a logging.Handler that delivers all records to host_log.

logging.basicConfig is changed to use logging.HostHandler by default.

Since any call like logging.warn(...) sets up logging via logging.basicConfig, all log messages are now delivered to crate::host_log, which will send them to tracing::event!.