angreal
Module utils

Filesystem utilities

Functions

check_up_to_date

context_to_map

get_task_files

Get a list of task files in given a path

is_angreal_project

Tests whether or not a current path is an angreal project

load_python

Loads a python file as a pyo3 PyModule

register

Registers the Command and Arg structs to the python api in the angreal module

render_dir

render_directory

repl_context_from_toml